requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
backports.entry-points-selectable==1.1.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2021.10.8
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.8
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.0.3
coloredlogs==15.0.1
cryptography==41.0.5
dataclasses-json==0.5.9
decorator==5.1.1
distlib==0.3.4
distro==1.9.0
dnspython==1.16.0
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.4.0
firebase-admin==6.2.0
Flask==2.0.2
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.0
httpx==0.27.0
humanfriendly==10.0
idna==3.3
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
langchain==0.0.231
langchainplus-sdk==0.0.20
langchain_openai
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==1.23.6
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==2.4.0
posthog==3.0.1
proto-plus==1.22.3
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
pymongo==3.12.0
Pympler==1.0.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyPDF2
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
reportlab==4.2.0
requests==2.28.1
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
twilio==7.3.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
virtualenv==20.10.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.0.2
yarl==1.9.2
zipp==3.16.0
pdf2image
langchain_pinecone
boto3
python-multipart
